Refactor integration tests
Split integration tests into separate files.
